Added shared-state demo. For some reason it is less efficient at this point, need to investigate.
git-svn-id: https://timeseriesdb.googlecode.com/svn/trunk@120 537f4964-0365-11de-bcc6-d9108be23e13

+++ TimeSeriesDb.Examples/DemoCompressed.cs
@@ -26,29 +26,42 @@ using System;
 using System.Collections.Generic;
 using System.IO;
 using System.Linq;
+using NYurik.TimeSeriesDb.Serializers.BlockSerializer;
 
 namespace NYurik.TimeSeriesDb.Examples
 {
-    internal static class DemoBinSeriesFile
+    internal class DemoCompressed : IExample
     {
-        public static void Run()
+        #region IExample Members
+
+        public void Run()
         {
-            const string filename = "BinSeriesFile.bts";
+            string filename = GetType().Name + ".bts";
+            if (File.Exists(filename)) File.Delete(filename);
 
-            Console.WriteLine("\n **** BinSeriesFile<long, ItemLngDbl> example ****\n");
-
-            if (File.Exists(filename))
-                File.Delete(filename);
-
-            // Create new BinSeriesFile file that stores a sequence of ItemLngDbl structs
+            // Create new BinCompressedSeriesFile file that stores a sequence of ItemLngDbl structs
             // The file is indexed by a long value inside ItemLngDbl marked with the [Index] attribute.
-            using (var bf = new BinSeriesFile<long, ItemLngDbl>(filename))
+            using (var bf = new BinCompressedSeriesFile<long, ItemLngDbl>(filename))
             {
                 //
                 // Initialize new file parameters and create it
                 //
                 bf.UniqueIndexes = true; // enforce index uniqueness
                 bf.Tag = "Sample Data"; // optionally provide a tag to store in the file header
+
+                //
+                // Configure value storage. This is the only difference with using BinSeriesFile.
+                //
+                // When a new instance of BinCompressedSeriesFile is created,
+                // RootField will be pre-populated with default configuration objects.
+                // Some fields, such as doubles, require additional configuration before the file can be initialized.
+                //
+                var root = (ComplexField) bf.FieldSerializer.RootField;
+
+                // This double will contain values with no more than 2 digits after the decimal points.
+                // Before serializing, multiply the value by 100 to convert to long.
+                ((ScaledDeltaField) root["Value"].Field).Multiplier = 100;
+
                 bf.InitializeNewFile(); // Finish new file initialization and create an empty file
 
 
@@ -75,7 +88,7 @@ namespace NYurik.TimeSeriesDb.Examples
             }
 
             // Re-open the file, allowing data modifications
-            // IWritableFeed<,> interface is better as it will work with compressed files as well
+            // IWritableFeed<,> interface is better as it will work with non-compressed files as well
             using (var bf = (IWritableFeed<long, ItemLngDbl>) BinaryFile.Open(filename, true))
             {
                 // Append a few more items with different ItemLngDbl.Value to tell them appart
@@ -92,7 +105,7 @@ namespace NYurik.TimeSeriesDb.Examples
             }
 
             // Re-open the file for reading only (file can be opened for reading in parallel, but only one write)
-            // IEnumerableFeed<,> interface is better as it will work with compressed files as well
+            // IEnumerableFeed<,> interface is better as it will work with non-compressed files as well
             using (var bf = (IEnumerableFeed<long, ItemLngDbl>) BinaryFile.Open(filename, true))
             {
                 // Show first item with index >= 5
@@ -119,5 +132,7 @@ namespace NYurik.TimeSeriesDb.Examples
             // cleanup
             File.Delete(filename);
         }
+
+        #endregion
     }
 }
